+++ engine/commands/publish.js
@@ -42,7 +42,8 @@ PublishOmen.prototype.run = function (filename) {
     var properties = [
         {
             name: 'Password',
-            hidden: true
+            hidden: true,
+            required: true
         }
     ];