+++ engine/utils/project_utils.js
@@ -17,6 +17,7 @@ var unirest = require('unirest'),
     tar = require('tar'),
     zlib = require('zlib'),
     Download = require('download'),
+    Handlebars = require('handlebars'),
     Q = require("q");
 
 /**
@@ -483,6 +484,36 @@ ProjectUtils.unpublishVersion = function (project, version, promptResult) {
         });
 
     return deferred.promise;
+};
+
+/**
+ * Generates the appserver configuration elements.
+ *
+ * @param {Project} project The project for which the appserver is created.
+ * @param {Object} result The prompt result.
+ */
+ProjectUtils.ubrokerTemplate = function (project, result) {
+    if (!fs.existsSync(path.resolve('./appserver'))) {
+        fs.mkdirSync(path.resolve('./appserver'));
+        fs.mkdirSync(path.resolve('./appserver/logs'));
+    }
+
+    var appserver = fs.readFileSync(path.resolve(__dirname + '/utilsTemplates/appserver.hbs'), "utf-8"),
+        appserverHbs = Handlebars.compile(appserver, {noEscape: true});
+
+    var output = appserverHbs({
+        name: project.get('name'),
+        propath: (OmenAPI.propath(project) + ';@{PROPATH}').replace(/;;/gi, ';'),
+        brokerLog: path.resolve('./appserver/logs/' + project.get('name') + '.broker.log'),
+        serverLog: path.resolve('./appserver/logs/' + project.get('name') + '.server.log'),
+        workdir: path.resolve('./appserver'),
+        port: result.Port,
+        opMode: result.opMode
+    });
+    console.log(output);
+
+    fs.writeFileSync(path.resolve("./appserver/.appserver"), output, "utf-8");
+    fs.writeFileSync(path.resolve("./appserver/appserver.pf"), '# appserver parameter file\n', "utf-8");
 };
 
 module.exports = ProjectUtils;